pause menu finished (+other)
The pause menu now works. I also made it so that the delta time variable is equal to 0 while the game is paused, so that all the different countdowns stop counting down. Furthermore, i fixed the issue where the players eyes disappear while directional keys are held in the pause menu. The main menu button works. I put a translucent overalay while the pause menu is on screen. Also, i found an out of bounds glitch involving damage knockback, which is now fixed.