some stuff
fixed some issues relating dash and knockback speeds not accounting for solid hitboxes. However a new issue has arisen now relating to the jellyfish hitbox, when and only when the one in room 3 spawns in it for some reason detect it is inside a wall and teleports out. should be an easy fix